If you want to change the direction wheels move when the stick is pressed, you modify the value being sent into the subsystem. You don't modify the interface...
There are other commands written and tested on the hardware authored to use the interface as documented.

It also makes no sense that neither wheel is inverted. I set the inversions to match what was tested on the robot Monday night, right inverted. Unless wiring was changed, that should still be correct.