Update Lift.java
Added methods to latch in the fully raised/fully lowered sensors in the event that we overshoot the sensors (possible at least for the raised sensor). This should prevent us from "breaking through" the magnetic field and continuing on our merry way into the wall. 
Similar to a dog and electric fence. If they have enough momentum their going to reach the other side of the invisible barrier... after that they aren't coming back.

I didn't use the encoder position as a method for tracking as I don't trust that the relative encoder will be a reliable indication of position through all modes of operation. Using command should be pretty fool proof unless someone commands through the barrier, momentarily commands in the direction opposite of travel (clearing the latch) then continues on away from the sensor.  On top of the software defense we're also going to add some more magnets (vertically stacked) so that the sensing region is increased.